@@ -11,47 +11,44 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~8.2.8",
-    "@angular/cdk": "^8.2.1",
-    "@angular/common": "~8.2.8",
-    "@angular/compiler": "~8.2.8",
-    "@angular/core": "~8.2.8",
-    "@angular/flex-layout": "^8.0.0-beta.27",
-    "@angular/forms": "~8.2.8",
-    "@angular/material": "^8.2.1",
-    "@angular/platform-browser": "~8.2.8",
-    "@angular/platform-browser-dynamic": "~8.2.8",
-    "@angular/pwa": "^0.803.12",
-    "@angular/router": "~8.2.8",
-    "@angular/service-worker": "~8.2.8",
-    "@zxing/ngx-scanner": "^2.0.1",
+    "@angular/animations": "~9.1.4",
+    "@angular/cdk": "^9.2.2",
+    "@angular/common": "~9.1.4",
+    "@angular/compiler": "~9.1.4",
+    "@angular/core": "~9.1.4",
+    "@angular/flex-layout": "^9.0.0-beta.29",
+    "@angular/forms": "~9.1.4",
+    "@angular/material": "^9.2.2",
+    "@angular/platform-browser": "~9.1.4",
+    "@angular/platform-browser-dynamic": "~9.1.4",
+    "@angular/router": "~9.1.4",
+    "@angular/service-worker": "~9.1.4",
+    "@zxing/ngx-scanner": "^3.0.0",
     "file-saver": "^2.0.2",
-    "hammerjs": "^2.0.8",
-    "ngx-bottom-nav": "^0.1.2",
-    "rxjs": "~6.5.3",
+    "rxjs": "~6.5.4",
     "tslib": "^1.10.0",
-    "xlsx": "^0.15.1",
+    "xlsx": "^0.16.0",
     "zone.js": "~0.10.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.803.6",
-    "@angular/cli": "^8.3.6",
-    "@angular/compiler-cli": "~8.2.8",
-    "@angular/language-service": "~8.2.8",
-    "@types/jasmine": "~3.4.1",
-    "@types/jasminewd2": "~2.0.7",
-    "@types/node": "~12.7.8",
+    "@angular-devkit/build-angular": "~0.901.4",
+    "@angular/cli": "~9.1.4",
+    "@angular/compiler-cli": "~9.1.4",
+    "@angular/language-service": "~9.1.4",
+    "@types/node": "^12.11.1",
+    "@types/jasmine": "~3.5.0",
+    "@types/jasminewd2": "~2.0.3",
     "codelyzer": "^5.1.2",
     "jasmine-core": "~3.5.0",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~4.3.0",
+    "karma": "~5.0.0",
     "karma-chrome-launcher": "~3.1.0",
     "karma-coverage-istanbul-reporter": "~2.1.0",
-    "karma-jasmine": "~2.0.1",
+    "karma-jasmine": "~3.0.1",
     "karma-jasmine-html-reporter": "^1.4.2",
-    "protractor": "~5.4.2",
-    "ts-node": "~8.4.1",
-    "tslint": "~5.20.0",
-    "typescript": "~3.4.3"
+    "protractor": "~5.4.3",
+    "ts-node": "~8.3.0",
+    "tslint": "~6.1.0",
+    "typescript": "~3.8.3"
   }
 }
