@@ -1,10 +1,19 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
-import { DashboardComponent } from './dashboard/dashboard.component';
+import { AdminComponent } from './admin.component';
+
 
 const routes: Routes = [
-  { path: '', redirectTo: 'dashboard', pathMatch: 'full' },
-  { path: "dashboard", component: DashboardComponent },
+  { 
+    path: "",
+    component: AdminComponent,
+    /*children: [
+      { path: '', redirectTo: 'dashboard', pathMatch: 'full' },
+      { path: "dashboard", component: DashboardComponent },
+      { path: "scan", component: ScanComponent },
+      { path: "settings", component: SettingsComponent }
+    ]*/
+  },
 ];
 
 @NgModule({
