@@ -1,14 +1,15 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-
+import { SharedModule } from '../shared/shared.module';
 import { AdminRoutingModule } from './admin-routing.module';
+import { AdminComponent } from './admin.component';
 import { DashboardComponent } from './dashboard/dashboard.component';
 
-
 @NgModule({
-  declarations: [DashboardComponent],
+  declarations: [AdminComponent, DashboardComponent],
   imports: [
     CommonModule,
+    SharedModule,
     AdminRoutingModule
   ]
 })
