@@ -1,18 +1,16 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-
+import { SharedModule } from '../shared/shared.module';
 import { AuthRoutingModule } from './auth-routing.module';
 import { LoginComponent } from './login/login.component';
-
-import { SharedModule } from '../shared/shared.module';
 import { LoginAdminComponent } from './login-admin/login-admin.component';
 
 @NgModule({
   declarations: [LoginComponent, LoginAdminComponent],
   imports: [
     CommonModule,
-    AuthRoutingModule,
-    SharedModule
+    SharedModule,
+    AuthRoutingModule
   ]
 })
 export class AuthModule { }
