@@ -1,8 +1,9 @@
 import { Component, OnInit } from '@angular/core';
-import { FormBuilder, FormGroup, Validators } from '@angular/forms';
 import { Router } from '@angular/router';
+import { FormBuilder, FormGroup, Validators } from '@angular/forms';
 import { MatSnackBar } from '@angular/material/snack-bar';
 import { PersonsService } from 'src/app/core/services/persons.service';
+import { UserLogged } from 'src/app/core/models/interfaces/user-logged';
 
 @Component({
   selector: 'app-login',
@@ -11,7 +12,6 @@ import { PersonsService } from 'src/app/core/services/persons.service';
 })
 export class LoginComponent implements OnInit {
 
-  user: any;
   signInForm: FormGroup;
   hidePassword = true;
   bSignIn = false;
@@ -21,22 +21,18 @@ export class LoginComponent implements OnInit {
     public snackBar: MatSnackBar,
     private formBuilder: FormBuilder,
     private personsService: PersonsService
-  ){}
+  ) { }
 
-  ngOnInit() {
-    this.InstantiateForm();
-  }
-
-  InstantiateForm(){
+  ngOnInit(): void {
     this.signInForm = this.formBuilder.group({
-      Identifier: ['', Validators.required],
+      Identifier: ['', [Validators.required, Validators.pattern("^[0-9]*$")]],
       Password: ['', Validators.required]
     });
   }
 
   openSnackBar(message: string){
     this.snackBar.open(message, 'Close', {
-      duration: 3000,
+      duration: 5000,
     });
   }
 
@@ -46,37 +42,34 @@ export class LoginComponent implements OnInit {
 
   handleSignIn(){
     this.bSignIn = true;
+    let credentialsFormData = new FormData();
+    credentialsFormData.append('Identifier', this.signInForm.get('Identifier').value);
+    credentialsFormData.append('Password', this.signInForm.get('Password').value);
 
-    let formData = new FormData();
-    formData.append('Identifier', this.signInForm.get('Identifier').value);
-    formData.append('Password', this.signInForm.get('Password').value);
-
-    this.personsService.validateUserCredentials(formData)
+    this.personsService.validateUserCredentials(credentialsFormData)
     .subscribe(
       res => {
         this.bSignIn = false;
-        let auxRes: any = res;
-
-        if(auxRes.type == 'error'){
-          this.openSnackBar('Invalid User or Password...');
+        /*Shocklogic's API returns 200 for ALL responses (a very very bad practice) 
+        so the way we can 'check' if the login was successful 
+        is with the TYPE attribute of the endpoint response*/
+        if(res.type == 'error'){
+          this.openSnackBar('Invalid User ID or Password...');
           return;
         }
-        
-        else if(auxRes.type == 'success'){
-          let auxUser = {
-            personId: auxRes.person_id,
-            clientId: auxRes.client_id,
-            projectId: auxRes.project_id
-          }
-          localStorage.setItem('leadLogged', JSON.stringify(auxUser));
+        else if(res.type == 'success'){
+          delete res.type; delete res.check; delete res.passed;
+          let userLogged: UserLogged = res;
+          localStorage.setItem('userLogged', JSON.stringify(userLogged));
           this.goToDashboard();
         }
       },
       err => {
+        console.log(err);
         this.bSignIn = false;
         this.openSnackBar('Something went wrong...');
       }
     );
   }
- 
+
 }
