@@ -3,10 +3,14 @@ import { TestBed } from '@angular/core/testing';
 import { ApiService } from './api.service';
 
 describe('ApiService', () => {
-  beforeEach(() => TestBed.configureTestingModule({}));
+  let service: ApiService;
+
+  beforeEach(() => {
+    TestBed.configureTestingModule({});
+    service = TestBed.inject(ApiService);
+  });
 
   it('should be created', () => {
-    const service: ApiService = TestBed.get(ApiService);
     expect(service).toBeTruthy();
   });
 });
