@@ -3,10 +3,14 @@ import { TestBed } from '@angular/core/testing';
 import { DeviceScansService } from './device-scans.service';
 
 describe('DeviceScansService', () => {
-  beforeEach(() => TestBed.configureTestingModule({}));
+  let service: DeviceScansService;
+
+  beforeEach(() => {
+    TestBed.configureTestingModule({});
+    service = TestBed.inject(DeviceScansService);
+  });
 
   it('should be created', () => {
-    const service: DeviceScansService = TestBed.get(DeviceScansService);
     expect(service).toBeTruthy();
   });
 });
