@@ -3,10 +3,14 @@ import { TestBed } from '@angular/core/testing';
 import { PersonsService } from './persons.service';
 
 describe('PersonsService', () => {
-  beforeEach(() => TestBed.configureTestingModule({}));
+  let service: PersonsService;
+
+  beforeEach(() => {
+    TestBed.configureTestingModule({});
+    service = TestBed.inject(PersonsService);
+  });
 
   it('should be created', () => {
-    const service: PersonsService = TestBed.get(PersonsService);
     expect(service).toBeTruthy();
   });
 });
