@@ -1,19 +1,22 @@
 import { Injectable } from '@angular/core';
 import { ApiService } from './api.service';
 
-const ENDPOINT_NAME = 'Persons';
-
 @Injectable({
   providedIn: 'root'
 })
-export class PersonsService extends ApiService {
+export class PersonsService {
+  private readonly personsUrl;
 
-  validateUserCredentials(user: any){                                /*login/Type/pid*/
-    return this.http.post(this.API_URL+'NONE-NONE-2-/'+ENDPOINT_NAME+'/login/', user, this.httpOptions);
+  constructor(private http: ApiService) {
+    this.personsUrl = 'Persons';
   }
 
-  getSpecificPersonRecord(clientId: string, projectId: string, personId: string){
-    return this.http.get(this.API_URL+clientId+'-'+projectId+'-2-/'+ENDPOINT_NAME+'/'+personId, this.httpOptions);
+  validateUserCredentials(userCredentials: FormData) :any {     /*login/Type/pid*/
+    return this.http.post('NONE-NONE-2-/'+this.personsUrl+'/login/', userCredentials, this.http.httpOptions);
+  }
+
+  getSpecificPersonRecord(clientId: string, projectId: string, personId: number){
+    return this.http.get(clientId+'-'+projectId+'-2-/'+this.personsUrl+'/'+personId, this.http.httpOptions);
   }
 
 }
