@@ -3,10 +3,14 @@ import { TestBed } from '@angular/core/testing';
 import { UsersService } from './users.service';
 
 describe('UsersService', () => {
-  beforeEach(() => TestBed.configureTestingModule({}));
+  let service: UsersService;
+
+  beforeEach(() => {
+    TestBed.configureTestingModule({});
+    service = TestBed.inject(UsersService);
+  });
 
   it('should be created', () => {
-    const service: UsersService = TestBed.get(UsersService);
     expect(service).toBeTruthy();
   });
 });
