@@ -1,7 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
-import { MatSnackBar } from '@angular/material/snack-bar';
-import { PersonsService } from 'src/app/core/services/persons.service';
+import { UserLogged } from 'src/app/core/models/interfaces/user-logged';
 
 @Component({
   selector: 'app-account',
@@ -10,50 +9,19 @@ import { PersonsService } from 'src/app/core/services/persons.service';
 })
 export class AccountComponent implements OnInit {
 
-  user: any;
-  userData: any;
-  //This handles the mat-progress-bar
-  bUser = true;
-  //This handles the error response when loading projects
-  bError = false;
+  userLogged: UserLogged;
 
   constructor(
     private router: Router,
-    public snackBar: MatSnackBar,
-    private personsService: PersonsService
   ) { }
 
-  ngOnInit() {
-    this.user = JSON.parse(localStorage.getItem('leadLogged'));
-    this.getUserData(this.user.clientId, this.user.projectId, this.user.personId);
+  ngOnInit(): void {
+    this.userLogged = JSON.parse(localStorage.getItem('userLogged'));
+    console.log(this.userLogged);
   }
 
-  goToLogin(){
-    this.router.navigate(['']);
-  }
-
-  goToSettings(){
+  goToHome(){
     this.router.navigate(['pages/settings']);
-  }
-
-  getUserData(clientId: string, projectId: string, personId: string){
-    this.personsService.getSpecificPersonRecord(clientId, projectId, personId)
-    .subscribe(
-      res => {
-        this.userData = res;
-        this.bUser = false;
-      },
-      err => {
-        this.bUser = false;
-        this.bError = true;
-      }
-    );
-  }
-
-  loadInfo(){
-    this.bUser = true;
-    this.bError = false;
-    this.getUserData(this.user.clientId, this.user.projectId, this.user.personId);
   }
 
 }
