@@ -1,58 +1,56 @@
-import { Component, OnInit, ViewChild } from '@angular/core';
-import {animate, state, style, transition, trigger} from '@angular/animations';
-import {MatPaginator} from '@angular/material/paginator';
-import {MatTableDataSource} from '@angular/material/table';
-import {MatSort} from '@angular/material/sort';
+import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
 import { MatSnackBar } from '@angular/material/snack-bar';
-import { PersonsService } from '../../core/services/persons.service';
-import { DeviceScansService } from '../../core/services/device-scans.service';
 import { map } from 'rxjs/operators';
+import { UserLogged } from 'src/app/core/models/interfaces/user-logged';
+import { DeviceScansService } from 'src/app/core/services/device-scans.service';
 
 @Component({
   selector: 'app-dashboard',
   templateUrl: './dashboard.component.html',
-  styleUrls: ['./dashboard.component.scss'],
-  animations: [
-    trigger('detailExpand', [
-      state('collapsed', style({height: '0px', minHeight: '0'})),
-      state('expanded', style({height: '*'})),
-      transition('expanded <=> collapsed', animate('225ms cubic-bezier(0.4, 0.0, 0.2, 1)')),
-    ]),
-  ],
+  styleUrls: ['./dashboard.component.scss']
 })
 export class DashboardComponent implements OnInit {
 
-  user: any;
-  alluserData: any;
-  contactsData: any;
+  userLogged: UserLogged;
+  myScansData: any;
+  filteredScansData: any[] = [];
   //This handles the mat-progress-bar
   bLeads = true;
-  //This handles the error response when loading projects
+  //This handles the error response when loading all the scans data
   bError = false;
-  //This handles the mat-spinner of the details button
-  bDetails = false;
   //The message for the custom error
   customError: string;
 
-  @ViewChild(MatPaginator, {static: true}) paginator: MatPaginator;
-  @ViewChild(MatSort, {static: true}) sort: MatSort;
+  _listFilter = '';
 
-  dataSource: any;
-  columnsToDisplay = ['Family_Name', 'First_Name', 'Company'];
-  expandedElement: any | null;
+  get listFilter(): string {
+    return this._listFilter;
+  }
+
+  set listFilter(value: string) {
+    this._listFilter = value;
+    this.filteredScansData = this.listFilter ? this.doFilter(this.listFilter) : this.myScansData;
+  }
+
+  doFilter(filterBy: string): any[] {
+    filterBy = filterBy.toLocaleLowerCase();
+    return this.myScansData.filter((participant: any) =>
+    participant.First_Name.toLocaleLowerCase().indexOf(filterBy) !== -1 ||
+    participant.Family_Name.toLocaleLowerCase().indexOf(filterBy) !== -1 ||
+    participant.Company.toLocaleLowerCase().indexOf(filterBy) !== -1
+    );
+  }
 
   constructor(
     private router: Router,
     public snackBar: MatSnackBar,
-    private personsService: PersonsService,
     private deviceScansService: DeviceScansService
   ) { }
 
-  ngOnInit() {
-    this.user = JSON.parse(localStorage.getItem('leadLogged'));
-    //this.getUserData(this.user.clientId, this.user.projectId, this.user.personId);
-    this.getTheContacts(this.user.clientId, this.user.projectId, this.user.personId);
+  ngOnInit(): void {
+    this.userLogged = JSON.parse(localStorage.getItem('userLogged'));
+    this.getTheContacts(this.userLogged.client_id, this.userLogged.project_id, this.userLogged.person_id);
   }
 
   openSnackBar(message: string){
@@ -61,25 +59,10 @@ export class DashboardComponent implements OnInit {
     });
   }
 
-  /*getUserData(clientId: string, projectId: string, personId: string){
-    this.personsService.getSpecificPersonRecord(clientId, projectId, personId)
-    .subscribe(
-      res => {
-        this.alluserData = res;
-        //console.log(this.alluserData);
-        if(false){
-          this.openSnackBar('Welcome '+this.alluserData.First_Name+' '+this.alluserData.Family_Name);
-        }
-      },
-      err => {
-        //console.log(err);
-        this.openSnackBar(err.message);
-      }
-    );
-  }*/
-
-  getTheContacts(clientId: string, projectId: string, personId: string){
-    this.deviceScansService.getAllScans(clientId, projectId, personId).pipe(
+  getTheContacts(clientId: string, projectId: string, personId: number){
+    this.deviceScansService.getAllScans(clientId, projectId, personId)
+    //We use the pipe and map to obtain only the attributes we need from each Lead
+    .pipe(
       map(
         (resp: any) => { 
           return resp.map(({Person_Id, Family_Name, First_Name, Company, Job_Title, EMail, Mobile}) =>
@@ -89,12 +72,10 @@ export class DashboardComponent implements OnInit {
     )
     .subscribe(
       res => {
-        this.filterLeads(res);
-        console.log(this.contactsData);
-        this.dataSource = new MatTableDataSource(this.contactsData);
-        this.dataSource.paginator = this.paginator;
-        this.dataSource.sort = this.sort;
         this.bLeads = false;
+        //We filter to get a single scan per Lead (The last one was done)
+        this.filterLeads(res);
+        this.filteredScansData = this.myScansData;
       },
       err => {
         if(err.error.text){
@@ -106,44 +87,8 @@ export class DashboardComponent implements OnInit {
     );
   }
 
-  applyFilter(filterValue: string) {
-    this.dataSource.filter = filterValue.trim().toLowerCase();
-  }
-
-  goToDetails(element){
-    this.bDetails = true;
-    this.bLeads = true;
-    this.deviceScansService.getNotesForAPerson(this.user.clientId, this.user.projectId, element.Person_Id, this.user.personId)
-    .subscribe(
-      res => {
-        let aux: any = res;
-        if(aux != null){
-          element.notes = aux.Notes;
-        }
-        else{
-          element.notes = '';
-        }
-        this.bDetails = false;
-        this.bLeads = false;
-        localStorage.setItem('leadDetails', JSON.stringify(element));
-        this.router.navigate(['pages/lead-details']);
-      },
-      err => {
-        this.bDetails = false;
-        this.bLeads = false;
-        this.openSnackBar('Something went wrong...');
-      }
-    );
-  }
-
-  loadLeads(){
-    this.bLeads = true;
-    this.bError = false;
-    this.getTheContacts(this.user.clientId, this.user.projectId, this.user.personId);
-  }
-
   filterLeads(leads){
-    this.contactsData = Array.from(new Set(leads.map(s => s.Person_Id)))
+    this.myScansData = Array.from(new Set(leads.map(s => s.Person_Id)))
     .map(id => {
         return {
           Person_Id: id,
@@ -155,6 +100,16 @@ export class DashboardComponent implements OnInit {
           Mobile: leads.find(s => s.Person_Id === id).Mobile
         };
     });
+  }
+
+  loadLeads(){
+    this.bLeads = true;
+    this.bError = false;
+    this.getTheContacts(this.userLogged.client_id, this.userLogged.project_id, this.userLogged.person_id);
+  }
+
+  selectLead(personId){
+    this.router.navigate(['pages/lead-details', personId]);
   }
 
 }
