@@ -10,11 +10,24 @@ export class LoginGuard implements CanActivate {
   constructor(private router: Router){}
   
   canActivate(){
-    if(localStorage.getItem('leadLogged')){
+    //
+    if(localStorage.getItem('userLogged') && !localStorage.getItem('adminLogged')){
       this.router.navigate(['pages'], { replaceUrl: true });
       return false;
     }
+    //
+    else if(localStorage.getItem('adminLogged') && !localStorage.getItem('userLogged')){
+      this.router.navigate(['admin'], { replaceUrl: true });
+      return false;
+    }
+    //
+    else if(localStorage.getItem('userLogged') && localStorage.getItem('adminLogged')){
+      localStorage.clear();
+      return true;
+    }
+    //
     else{
+      localStorage.clear();
       return true;
     }
     
