@@ -3,54 +3,39 @@ import { CommonModule } from '@angular/common';
 
 import {MatFormFieldModule} from '@angular/material/form-field';
 import {MatInputModule} from '@angular/material/input';
-import {MatIconModule} from '@angular/material/icon';
-import {MatButtonModule} from '@angular/material/button';
-import {MatToolbarModule} from '@angular/material/toolbar';
-import {MatTableModule} from '@angular/material/table';
-import {MatMenuModule} from '@angular/material/menu';
-import {MatBadgeModule} from '@angular/material/badge';
-import {MatSnackBarModule} from '@angular/material/snack-bar';
-import {MatProgressSpinnerModule} from '@angular/material/progress-spinner';
-import {MatPaginatorModule} from '@angular/material/paginator';
-import {MatSortModule} from '@angular/material/sort';
-import {MatProgressBarModule} from '@angular/material/progress-bar';
 import {MatSelectModule} from '@angular/material/select';
-import {MatListModule} from '@angular/material/list';
-import {MatRippleModule} from '@angular/material/core';
+
+import {MatMenuModule} from '@angular/material/menu';
+import {MatSidenavModule} from '@angular/material/sidenav';
+import {MatToolbarModule} from '@angular/material/toolbar';
+
 import {MatCardModule} from '@angular/material/card';
-import {MatTabsModule} from '@angular/material/tabs';
-import {MatChipsModule} from '@angular/material/chips';
-import {MatGridListModule} from '@angular/material/grid-list';
-import {MatStepperModule} from '@angular/material/stepper';
-import {MatExpansionModule} from '@angular/material/expansion';
+import {MatListModule} from '@angular/material/list';
+
+import {MatButtonModule} from '@angular/material/button';
+import {MatIconModule} from '@angular/material/icon';
+import {MatProgressSpinnerModule} from '@angular/material/progress-spinner';
+import {MatProgressBarModule} from '@angular/material/progress-bar';
+
+import {MatSnackBarModule} from '@angular/material/snack-bar';
+
 
 const MATERIAL_MODULES = [
   CommonModule,
   MatFormFieldModule,
   MatInputModule,
-  MatIconModule,
-  MatButtonModule,
-  MatToolbarModule,
-  MatTableModule,
-  MatMenuModule,
-  MatBadgeModule,
-  MatSnackBarModule,
-  MatProgressSpinnerModule,
   MatSelectModule,
-  MatPaginatorModule,
-  MatSortModule,
-  MatProgressBarModule,
-  MatListModule,
-  MatRippleModule,
+  MatMenuModule,
+  MatSidenavModule,
+  MatToolbarModule,
   MatCardModule,
-  MatTabsModule,
-  MatChipsModule,
-  MatGridListModule,
-  MatStepperModule,
-  MatExpansionModule
+  MatListModule,
+  MatButtonModule,
+  MatIconModule,
+  MatProgressSpinnerModule,
+  MatProgressBarModule,
+  MatSnackBarModule
 ]
-
-
 
 @NgModule({
   declarations: [],
