@@ -1,26 +1,25 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { ReactiveFormsModule } from '@angular/forms';
+import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { MaterialModule } from './material/material.module';
 import { FlexLayoutModule } from "@angular/flex-layout";
-import {BottomNavModule} from 'ngx-bottom-nav';
 import { ZXingScannerModule } from '@zxing/ngx-scanner';
 
 @NgModule({
   declarations: [],
   imports: [
     CommonModule,
+    FormsModule,
     ReactiveFormsModule,
     MaterialModule,
     FlexLayoutModule,
-    BottomNavModule,
     ZXingScannerModule
   ],
   exports: [
+    FormsModule,
     ReactiveFormsModule,
     MaterialModule,
     FlexLayoutModule,
-    BottomNavModule,
     ZXingScannerModule
   ]
 })
